refactor(server): isolate account and connector grant views
