+++ services/server/src/app.ts
@@ -106,6 +106,7 @@ import {
   registerConnectorPairingRoutes
 } from "./features/connectors/pairing-routes.js";
 import { registerAccountSessionRoutes } from "./features/account/session-routes.js";
+import { registerAccountOverviewRoute } from "./features/account/me-routes.js";
 import { registerMirrorPairingRoutes } from "./features/mirrors/pairing-routes.js";
 import {
   registerConnectorManagementRoutes
@@ -191,6 +192,12 @@ import {
   createOrUpdateGrant,
   syncHostedNotificationGrant
 } from "./features/grants/service.js";
+import {
+  registerConnectorGrantRoutes
+} from "./features/grants/connector-routes.js";
+import {
+  liveAuthorizationCollections
+} from "./features/authorizations/local-collections.js";
 import { sessionToken } from "./platform/session-cookies.js";
 import { audit } from "./platform/audit-events.js";
 import {
@@ -473,328 +480,17 @@ export async function buildApp(options: BuildOptions) {
     hostedProvider: options.hostedProvider,
     allowInsecureManifests: options.allowInsecureManifests
   });
-
-  app.get("/v1/me", async (request, reply) => {
-    const authenticated = await requireUser(request, reply, options.db, options.tailscaleAuth);
-    if (!authenticated) return;
-    if (options.hostedCollections) {
-      await recoverExpiredAuthorityTransfers(options.db, options.hostedProvider, hostedReference);
-    }
-    const { authentication_provider: authenticationProvider, ...user } = authenticated;
-    const connectors = await options.db.query(
-      `SELECT c.id, c.name, c.last_seen_at, c.created_at
-       FROM connectors c
-       WHERE c.user_id = $1 AND c.revoked_at IS NULL
-       ORDER BY c.created_at`,
-      [user.id]
-    );
-    const localCatalog = await listLocalCollectionsVisibleToUser(
-      options.db,
-      user.id
-    );
-    const localAuthorityIds = localCatalog.map(
-      (collection) => collection.authorityRowId
-    );
-    const collections = localAuthorityIds.length
-      ? await options.db.query(
-          `SELECT col.id, col.connector_id, col.local_id, col.display_name,
-                  col.spec_version, col.enabled, col.authority_state,
-                  col.authority_epoch, col.contracts, col.last_seen_at,
-                  connector.name AS connector_name
-           FROM collections col
-           JOIN connectors connector ON connector.id = col.connector_id
-           WHERE col.id IN (${sqlPlaceholders(localAuthorityIds.length)})
-             AND connector.revoked_at IS NULL
-             AND col.authority_state = 'active'
-             AND col.present = true
-           ORDER BY col.display_name`,
-          localAuthorityIds
-        )
-      : { rows: [] };
-    const hostedCatalog = options.hostedCollections
-      ? (await listHostedCollectionsVisibleToUser(options.db, user.id))
-          .filter((collection) =>
-            collection.locator.authorityState !== "importing"
-          )
-      : [];
-    const hostedCollections = {
-      rows: await Promise.all(hostedCatalog.map(async (collection) => ({
-        id: collection.locator.collectionId,
-        display_name: collection.locator.displayName,
-        template: collection.template,
-        contracts: collection.contracts,
-        provider_url: collection.locator.providerUrl ?? null,
-        authority_state:
-          collection.locator.authorityState as "active" | "transferring" | "transferred",
-        authority_epoch: collection.locator.authorityEpoch,
-        transferred_collection_id: collection.transferredCollectionId,
-        created_at: collection.createdAt,
-        access: accessView(requireCollectionAction(
-          await resolveHostedCollectionAccess(
-            options.db,
-            user.id,
-            collection.locator.collectionId
-          ),
-          "collection.discover"
-        ))
-      })))
-    };
-    const hostedReplicas = {
-      rows: await hostedMirrorReplicas(
-        options.db,
-        hostedCollections.rows.map((collection) => collection.id)
-      )
-    };
-    const hostedReplicaStatuses = new Map<string, {
-      head: number;
-      acknowledged_sequence: number;
-      last_seen_at: string | null;
-      token_expires_at: string;
-    }>();
-    if (options.hostedProvider) {
-      const statusGroups = await Promise.all(hostedCollections.rows.map(async (collection) => {
-        try {
-          return await options.hostedProvider!.replicaStatuses(collection.id);
-        } catch (error) {
-          request.log.warn({ error, collection_id: collection.id }, "Hosted mirror status is unavailable");
-          return [];
-        }
-      }));
-      for (const status of statusGroups.flat()) hostedReplicaStatuses.set(status.id, status);
-    }
-    const grants = await options.db.query(
-      `SELECT g.id, g.operations, g.scope, g.created_at, g.revoked_at,
-              CASE WHEN g.application_origin = '' THEN a.homepage
-                   ELSE g.application_origin END AS application_origin,
-              COALESCE(col.local_id, g.hosted_collection_id) AS collection_id,
-              a.id AS application_id,
-              a.family_identity AS application_family_id,
-              a.distribution, a.name AS application_name,
-              a.homepage, a.project_url, a.icon,
-              COALESCE(col.display_name, hosted.display_name) AS collection_name,
-              CASE WHEN g.hosted_collection_id IS NULL THEN 'local' ELSE 'hosted' END AS collection_kind
-       FROM grants g
-       JOIN applications a ON a.id = g.application_id
-       LEFT JOIN collections col ON col.id = g.collection_id
-       LEFT JOIN hosted_collections hosted ON hosted.id = g.hosted_collection_id
-       WHERE g.user_id = $1
-         AND (g.activated_at IS NOT NULL OR g.revoked_at IS NOT NULL)
-       ORDER BY g.created_at DESC`,
-      [user.id]
-    );
-    const pendingAuthorizations = await options.db.query(
-      `SELECT ar.id, ar.flow, ar.user_code, ar.requested_operations,
-              ar.collection_id, ar.expires_at,
-              a.id AS application_id, a.distribution, a.name AS application_name,
-              a.homepage, a.project_url, a.icon,
-              a.requirements, a.provisions, a.notifications
-       FROM authorization_requests ar
-       JOIN applications a ON a.id = ar.application_id
-       WHERE ar.user_id = $1 AND ar.completed_at IS NULL AND ar.denied_at IS NULL
-         AND ar.expires_at > now()
-       ORDER BY ar.expires_at`,
-      [user.id]
-    );
-    const authenticationSettings = await authenticationPolicy.current();
-    return {
-      user,
-      hosted_collections_available: options.hostedCollections === true,
-      authentication: {
-        provider: authenticationProvider ?? (options.tailscaleAuth ? "tailscale" : "session"),
-        registration: authenticationSettings.registrationMode
-      },
-      connectors: connectors.rows,
-      collections: await Promise.all(collections.rows.map(async (collection) => {
-        const access = await resolveLocalCollectionAccess(
-          options.db,
-          user.id,
-          collection.id
-        );
-        return {
-          ...collection,
-          id: collection.local_id,
-          ...(access ? {
-            access: accessView(access),
-            authority: {
-              kind: "local",
-              collection_id: access.collection.collectionId,
-              epoch: access.collection.authorityEpoch,
-              state: access.collection.authorityState
-            }
-          } : {})
-        };
-      })),
-      hosted_collections: hostedCollections.rows.map((collection) => ({
-        ...collection,
-        provider_url: collection.provider_url ?? publicUrl,
-        spec_version: "0.3.0",
-        authority_epoch: Number(collection.authority_epoch),
-        authority: {
-          kind: "hosted",
-          collection_id: collection.id,
-          epoch: Number(collection.authority_epoch),
-          state: collection.authority_state
-        },
-        contracts: contractRequirements(effectiveHostedContractDescriptors(collection.contracts, collection.template)),
-        replicas: hostedReplicas.rows
-          .filter((replica) => replica.collection_id === collection.id)
-          .map((replica) => ({
-            ...replica,
-            sync_status: hostedReplicaStatuses.get(replica.id) ?? null
-          }))
-      })),
-      grants: grants.rows.map((grant) => ({
-        ...grant,
-        application_origin: normalizedApplicationOrigin(grant.application_origin)
-      })),
-      pending_authorizations: await Promise.all(pendingAuthorizations.rows.map(async (authorization) => {
-        const live = requiresHostedCollection(authorization.requirements)
-          ? { collections: [], unavailable_connectors: [] }
-          : await liveAuthorizationCollections(
-              options.db,
-              relay,
-              user.id,
-              authorization.id
-            );
-        return {
-          ...authorization,
-          available_collections: live.collections,
-          unavailable_connectors: live.unavailable_connectors
-        };
-      }))
-    };
+  registerAccountOverviewRoute(app, {
+    db: options.db,
+    relay,
+    publicUrl,
+    authenticationPolicy,
+    tailscaleAuth: options.tailscaleAuth,
+    hostedCollections: options.hostedCollections,
+    hostedProvider: options.hostedProvider,
+    hostedReference
   });
-
-  app.post("/v1/connectors/grants", async (request, reply) => {
-    const connector = await requireConnector(request, reply, options.db);
-    if (!connector) return;
-    const input = z.object({
-      application_id: z.uuid(),
-      collection_id: z.uuid(),
-      operations: z.array(operationSchema).min(1),
-      contracts: z.array(collectionContractDescriptorSchema).max(100).optional()
-    }).parse(request.body);
-    if (input.contracts) {
-      await options.db.query(
-        `UPDATE collections SET contracts = $3::jsonb, last_seen_at = now()
-         WHERE connector_id = $1 AND local_id = $2 AND enabled = true
-           AND present = true AND authority_state = 'active'`,
-        [connector.id, input.collection_id, JSON.stringify(input.contracts)]
-      );
-    }
-    const collection = await options.db.query<{ id: string; contracts: CollectionContractDescriptor[]; spec_version: string }>(
-      `SELECT id, contracts, spec_version FROM collections
-       WHERE connector_id = $1 AND local_id = $2 AND enabled = true
-         AND present = true AND authority_state = 'active'`,
-      [connector.id, input.collection_id]
-    );
-    if (!collection.rows[0]) return reply.code(404).send(apiError("collection_not_found", "Collection is not synchronized yet."));
-    const application = await options.db.query<{
-      id: string;
-      distribution: "web" | "portable";
-      homepage: string;
-      requirements: ApplicationRequirements;
-      notifications: ApplicationNotifications;
-    }>(
-      "SELECT id, distribution, homepage, requirements, notifications FROM applications WHERE id = $1",
-      [input.application_id]
-    );
-    if (!application.rows[0]) return reply.code(404).send(apiError("application_not_found", "Application not found."));
-    if (application.rows[0].distribution === "portable") {
-      return reply.code(409).send(apiError(
-        "portable_approval_required",
-        "Downloaded applications must use their key-bound device authorization request."
-      ));
-    }
-    if (requiresHostedCollection(application.rows[0].requirements)) {
-      return reply.code(409).send(apiError(
-        "incompatible_collection",
-        "This application requires an mdbase cloud collection."
-      ));
-    }
-    assertOperationsAllowedByRequirements(input.operations, application.rows[0].requirements);
-    assertCollectionSupportsOperations(collection.rows[0].spec_version, input.operations);
-    const scope = scopeForRequirements(
-      application.rows[0].requirements,
-      collection.rows[0].contracts
-    );
-    if (!contractsSatisfy(
-      collection.rows[0].contracts,
-      requiredContractsForRequirements(application.rows[0].requirements)
-    )) {
-      return reply.code(409).send(apiError(
-        "incompatible_collection",
-        "This collection does not provide the contracts required by the application."
-      ));
-    }
-    const grant = await createOrUpdateGrant(options.db, {
-      userId: connector.user_id,
-      applicationId: input.application_id,
-      collectionId: collection.rows[0].id,
-      operations: input.operations,
-      scope,
-      applicationOrigin: new URL(application.rows[0].homepage).origin,
-      notificationCriteria: application.rows[0].notifications.criteria
-    });
-    await relay.pushPolicy(connector.id);
-    await audit(options.db, connector.user_id, "grant.created", grant.id, {
-      ...input,
-      connector_id: connector.id
-    });
-    return reply.code(201).send({ grant });
-  });
-
-  app.patch("/v1/connectors/grants/:grantId", async (request, reply) => {
-    const connector = await requireConnector(request, reply, options.db);
-    if (!connector) return;
-    const { grantId } = z.object({ grantId: z.uuid() }).parse(request.params);
-    const input = z.object({ operations: z.array(operationSchema).min(1) }).parse(request.body);
-    const current = await options.db.query<{ requirements: ApplicationRequirements; spec_version: string }>(
-      `SELECT a.requirements, col.spec_version FROM grants g
-       JOIN applications a ON a.id = g.application_id
-       JOIN collections col ON col.id = g.collection_id
-       WHERE g.id = $1 AND g.revoked_at IS NULL AND g.activated_at IS NOT NULL
-         AND g.collection_id IN
-         (SELECT id FROM collections WHERE connector_id = $2)`,
-      [grantId, connector.id]
-    );
-    if (!current.rows[0]) return reply.code(404).send(apiError("grant_not_found", "Active grant not found."));
-    assertOperationsAllowedByRequirements(input.operations, current.rows[0].requirements);
-    assertCollectionSupportsOperations(current.rows[0].spec_version, input.operations);
-    const grant = await options.db.query(
-      `UPDATE grants SET operations = $3::jsonb
-       WHERE id = $1 AND revoked_at IS NULL AND activated_at IS NOT NULL
-         AND collection_id IN
-         (SELECT id FROM collections WHERE connector_id = $2)
-       RETURNING id, operations`,
-      [grantId, connector.id, JSON.stringify([...new Set(input.operations)])]
-    );
-    if (!grant.rows[0]) return reply.code(404).send(apiError("grant_not_found", "Active grant not found."));
-    await rotateGrantEncryption(options.db, grantId);
-    await relay.pushPolicy(connector.id);
-    await audit(options.db, connector.user_id, "grant.updated", grantId, input);
-    return { grant: grant.rows[0] };
-  });
-
-  app.delete("/v1/connectors/grants/:grantId", async (request, reply) => {
-    const connector = await requireConnector(request, reply, options.db);
-    if (!connector) return;
-    const { grantId } = z.object({ grantId: z.uuid() }).parse(request.params);
-    const active = await options.db.query(
-      `UPDATE grants SET revoked_at = now()
-       WHERE id = $1 AND revoked_at IS NULL AND activated_at IS NOT NULL
-         AND collection_id IN
-         (SELECT id FROM collections WHERE connector_id = $2)
-       RETURNING id`,
-      [grantId, connector.id]
-    );
-    if (!active.rows[0]) return reply.code(404).send(apiError("grant_not_found", "Active grant not found."));
-    await options.db.query("UPDATE access_tokens SET revoked_at = now() WHERE grant_id = $1", [grantId]);
-    await options.db.query("UPDATE refresh_tokens SET revoked_at = now() WHERE grant_id = $1", [grantId]);
-    await relay.pushPolicy(connector.id);
-    await audit(options.db, connector.user_id, "grant.revoked", grantId, { connector_id: connector.id });
-    return { ok: true };
-  });
+  registerConnectorGrantRoutes(app, { db: options.db, relay });
 
   app.post("/v1/connectors/authorization-requests/:requestId/approve", async (request, reply) => {
     const connector = await requireConnector(request, reply, options.db);
@@ -1656,174 +1352,6 @@ export async function buildApp(options: BuildOptions) {
   }
 
   return { app, relay };
-}
-
-interface LiveAuthorizationCollection {
-  id: string;
-  offer_id: string;
-  kind: "local";
-  connector_name: string;
-  display_name: string;
-  spec_version: string;
-  contracts: CollectionContractDescriptor[];
-  access: ReturnType<typeof accessView>;
-}
-
-async function liveAuthorizationCollections(
-  db: DatabasePool,
-  relay: RelayHub,
-  userId: string,
-  authorizationId: string
-): Promise<{
-  collections: LiveAuthorizationCollection[];
-  unavailable_connectors: Array<{
-    connector_id: string;
-    connector_name: string;
-    reason: "offline" | "paused";
-  }>;
-}> {
-  await db.query(
-    `DELETE FROM authorization_collection_offers
-     WHERE authorization_id = $1 AND expires_at <= now()`,
-    [authorizationId]
-  );
-  const visibleCollections = await listLocalCollectionsVisibleToUser(db, userId);
-  const visibleByConnector = new Map<string, Set<string>>();
-  for (const collection of visibleCollections) {
-    if (
-      collection.authorityState !== "active"
-      || !collection.connectorId
-    ) continue;
-    const ids = visibleByConnector.get(collection.connectorId) ?? new Set();
-    ids.add(collection.authorityRowId);
-    visibleByConnector.set(collection.connectorId, ids);
-  }
-  const ownerConnectors = await db.query<{
-    id: string;
-    name: string;
-    inventory_revision: string | number;
-  }>(
-    `SELECT id, name, inventory_revision FROM connectors
-     WHERE user_id = $1 AND revoked_at IS NULL
-     ORDER BY created_at`,
-    [userId]
-  );
-  const ownerConnectorIds = new Set(ownerConnectors.rows.map(({ id }) => id));
-  const sharedConnectorIds = [...visibleByConnector.keys()]
-    .filter((id) => !ownerConnectorIds.has(id));
-  const sharedConnectors = sharedConnectorIds.length
-    ? await db.query<{
-        id: string;
-        name: string;
-        inventory_revision: string | number;
-      }>(
-        `SELECT id, name, inventory_revision FROM connectors
-         WHERE id IN (${sqlPlaceholders(sharedConnectorIds.length)})
-           AND revoked_at IS NULL
-         ORDER BY created_at`,
-        sharedConnectorIds
-      )
-    : { rows: [] };
-  const connectors = {
-    rows: [...ownerConnectors.rows, ...sharedConnectors.rows]
-  };
-  const settled = await Promise.allSettled(connectors.rows.map(async (connector) => ({
-    connector,
-    response: await relay.authorizationOffers(connector.id, authorizationId)
-  })));
-  const collections: LiveAuthorizationCollection[] = [];
-  const unavailableConnectors: Array<{
-    connector_id: string;
-    connector_name: string;
-    reason: "offline" | "paused";
-  }> = [];
-
-  for (const [index, result] of settled.entries()) {
-    const connector = connectors.rows[index];
-    if (result.status === "rejected") {
-      unavailableConnectors.push({
-        connector_id: connector.id,
-        connector_name: connector.name,
-        reason: "offline"
-      });
-      continue;
-    }
-    if (result.value.response.paused) {
-      unavailableConnectors.push({
-        connector_id: connector.id,
-        connector_name: connector.name,
-        reason: "paused"
-      });
-      continue;
-    }
-    const authoritative = await db.query<{
-      id: string;
-      local_id: string;
-      authority_epoch: string | number;
-    }>(
-      `SELECT id, local_id, authority_epoch FROM collections
-       WHERE connector_id = $1
-         AND present = true AND enabled = true AND authority_state = 'active'`,
-      [connector.id]
-    );
-    const visibleIds = visibleByConnector.get(connector.id) ?? new Set();
-    const byLocalId = new Map(authoritative.rows
-      .filter((collection) => visibleIds.has(collection.id))
-      .map((collection) => [collection.local_id, collection] as const));
-    for (const offered of result.value.response.collections) {
-      const collection = byLocalId.get(offered.collection_id);
-      if (!collection) continue;
-      const offer = await db.query<{ id: string }>(
-        `INSERT INTO authorization_collection_offers
-           (id, authorization_id, user_id, connector_id, collection_id, local_id,
-            authority_epoch, inventory_revision, expires_at)
-         VALUES ($1, $2, $3, $4, $5, $6, $7, $8, now() + interval '45 seconds')
-         ON CONFLICT(authorization_id, connector_id, collection_id) DO UPDATE SET
-           local_id = excluded.local_id,
-           authority_epoch = excluded.authority_epoch,
-           inventory_revision = excluded.inventory_revision,
-           expires_at = excluded.expires_at
-         WHERE authorization_collection_offers.consumed_at IS NULL
-         RETURNING id`,
-        [
-          randomUUID(),
-          authorizationId,
-          userId,
-          connector.id,
-          collection.id,
-          offered.collection_id,
-          Number(collection.authority_epoch),
-          Number(connector.inventory_revision)
-        ]
-      );
-      if (!offer.rows[0]) continue;
-      const access = await resolveLocalCollectionAccess(
-        db,
-        userId,
-        collection.id
-      );
-      if (!access || !access.actions.has("application.authorize")) continue;
-      collections.push({
-        id: offered.collection_id,
-        offer_id: offer.rows[0].id,
-        kind: "local",
-        connector_name: connector.name,
-        display_name: offered.display_name,
-        spec_version: offered.spec_version,
-        contracts: offered.contracts,
-        access: accessView(access)
-      });
-    }
-  }
-
-  collections.sort((left, right) =>
-    left.display_name.localeCompare(right.display_name, undefined, { sensitivity: "base" })
-    || left.connector_name.localeCompare(right.connector_name, undefined, { sensitivity: "base" })
-  );
-  return {
-    collections,
-    unavailable_connectors: unavailableConnectors
-  };
 }
 
 async function approvePortalAuthorization(
